restricted pages

+++ studybud/base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -10,6 +12,10 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -29,9 +35,11 @@ def loginPage(request):
     context = {}
     return render(request, "base/login_register.html", context)
 
+
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -50,6 +58,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, "base/room.html", context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -62,9 +72,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -75,8 +90,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
